Initial commit.
This represents the state of the project as I left it on November 29, 2023. You may notice that the code is written in English but the documentation (outside of the code) is in French. This is because I was more comfortable with French than with English. This French documentation was not complete, so I will reread the code and add what is missing. After that, I will produce a final documentation in English.